Move holoviews extension loader to its own module
